optimized the code structure
