+++ modules/globalGameAttributes.py
@@ -1,0 +1,12 @@
+from tkinter import *
+
+def init():
+    global root
+    global frame1
+    global which_potion
+
+    root = Tk()
+    frame1 = ""
+
+    # variable that lets you select the potion that you want to take.
+    which_potion = 0
