+++ modules/player.py
@@ -1,0 +1,21 @@
+gold = 1500  # used in store to buy items
+hp = 100  # user's hp
+opp_hp = 100  # monster's hp
+extra_hp = 0  # extra opponent hp for boss round
+Iron_Sword = False  # increases attack by 20
+Mythril_Sword = False  # increases attack by 30
+Orichalium_Sword = False  # increases attack by 40
+Iron_Armour = False  # decreases opp_att by 10
+Mythril_Armour = False  # decreases opp_att by 20
+Orichalium_Armour = False  # decreases opp_att by 30
+
+BunSamosa_Armour= False #decreases opp_att by 50
+ACM_Armour= False #decreases opp_att by 60
+Jade_Armour = False # increases hp by 10
+Diamond_Armour = False #increases hp by 20
+Silver_Armour = False #decreaases Opp_ATT by 40
+Gold_Armour= False #decreases opp_att by 70
+
+potion = 1  # increases hp by 30. Cost=300 gold
+ultra_potion = 1  # increases hp by 50. Cost=600 gold
+medium_potion = 1  # increases hp by 40, cost=450 gold
